before trying quantized

+++ temp/MicroFlow_Q.cpp
@@ -1,7 +1,35 @@
 #include "Arduino.h"
 #include "MicroFlow_Q.h"
 
-MicroMLP::MicroMLP(int la, int* top, int* w, int* b, int* a){
+
+// bool matrix_multiply(int R, int R2, int C2, double* a, double* b, double* result, int size){
+//  // double result[R2];
+//   double cpy[R];
+//   for (int i=0;i<R;i++){
+//     cpy[i] = a[i];
+//   }
+//   int columnB = 0;
+//   for (int i = 0; i < R2; i++) {
+//     int bIter = columnB;
+//     double sum = 0;
+//     for (int j = 0; j < R; j++) {
+//       //printf("==%f %f==", a[j], b[bIter]);
+//       sum += cpy[j] * b[bIter];
+//       bIter += R2;
+//     }
+//     //printf("\n");
+//     result[i] = sum;
+//     if (i < size) {
+//     } else {
+//       return false;
+//     }
+//     columnB ++;
+//   }
+//   return true;
+//   // return result;
+// }
+
+MicroMLP::MicroMLP(int la, int* top, double* w, double* b, int* a){
   layers = la;
   topology = top;
   weights = w;
@@ -9,7 +37,7 @@ MicroMLP::MicroMLP(int la, int* top, int* w, int* b, int* a){
   activations = a;
 }
 
-MicroMLP::MicroMLP(int la, int* top, int* w, int* b, int a){
+MicroMLP::MicroMLP(int la, int* top, double* w, double* b, int a){
   layers = la;
   topology = top;
   weights = w;
@@ -18,7 +46,7 @@ MicroMLP::MicroMLP(int la, int* top, int* w, int* b, int a){
   allSameActiv = true;
 }
 
-void activate(int l, int* z, int activation) {
+void activate(int l, double* z, int activation) {
   for (int i = 0; i < l; i++) {
     if (activation == SIGMOID) {
       z[i] = 1 / (1 + exp(-z[i]));
@@ -34,7 +62,7 @@ void activate(int l, int* z, int activation) {
   }
 }
 
-void MicroMLP::feedforward(int* input, int* out){
+void MicroMLP::feedforward(double* input, double* out){
   int maxLayer = 0;
   for (int i=0;i<layers;i++){
     if (topology[i] > maxLayer){
@@ -42,7 +70,7 @@ void MicroMLP::feedforward(int* input, int* out){
     }
   }
 
-  int x[maxLayer];
+  double x[maxLayer];
   for (int i=0;i<topology[0];i++){
     x[i] = input[i];
   }
@@ -51,14 +79,14 @@ void MicroMLP::feedforward(int* input, int* out){
   for (int l=0;l<layers-1;l++){
 
     //Matrix----
-    int cpy[topology[l]];
+    double cpy[topology[l]];
     for (int i=0;i<topology[l];i++){
       cpy[i] = x[i];
     }
     int columnB = 0;
     for (int i = 0; i < topology[l+1]; i++) {
       int bi = columnB;
-      int sum = 0;
+      double sum = 0;
       
       for (int j = 0; j < topology[l]; j++) {
         sum += cpy[j] * weights[bi+weightAdder];
@@ -67,6 +95,7 @@ void MicroMLP::feedforward(int* input, int* out){
       x[i] = sum;
       columnB ++;
     }
+    //matrix_multiply(topology[l], topology[l+1], topology[l], x, weightStack, x, topology[l+1]);
     
     for (int i=0;i<topology[l+1];i++){
       x[i] += biases[i+biasAdder];
